fix(UpsertInterventionsJobConfiguration): Add error handling (#119)
**WHAT**

We've wrapped the .run() command for the UpsertInterventionJob in a
try/catch block

**WHY**

Pre-prod has found itself in a failing crashing loop because it started,
and has not finished, an UpsertInterventionsJob.

This is causing the start-job to fail.  Our Kotlin server can't handle this error, and so our server is unable to stand up.

This fix (likely 1/2) fixes the immediate problem, and allows the server
to handle a failure/error in the running of this job.

**WHAT NEXT**

We need to clear out the hanging jobs, so that we can actually run the
job in the future if this reoccurs.