APG-1170: Add documentation around the ReferralDetailsDto
**WHAT**

This is a sibling commit to a change made in the manage-and-deliver
service, which introduces some better documentation and annotations to
code around the Referrals endpoints.

**WHY**

We are about to start using the sourced_from data (license or
requirement) in the manage-and-deliver service.  This PR is a bit of a
tidy-up alongside using that, to make sure it's clear what is being sent
over, and what the data represents.