APG-1170: Add documentation around the ReferralDetailsDto (#241)
* APG-1170: Add documentation around the ReferralDetailsDto

**WHAT**

This is a sibling commit to a change made in the manage-and-deliver
service, which introduces some better documentation and annotations to
code around the Referrals endpoints.

**WHY**

We are about to start using the sourced_from data (license or
requirement) in the manage-and-deliver service.  This PR is a bit of a
tidy-up alongside using that, to make sure it's clear what is being sent
over, and what the data represents.

* APG-1170(tests): Update tests to accommodate Referral::sourced_from

**WHAT**

This commit introduces changes to a number of test files related to the
fetching and creation of a ReferralEntity from the Find&Refer API.

It ensures that the relevant fields are being properly mocked out, with
realistic values, and that those values are feeding through into the
creation of the ReferralEntity.